+++ geofileops/_compat.py
@@ -21,6 +21,7 @@ GDAL_GTE_39 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.9")
 GDAL_GTE_3101 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.10.1")
 GDAL_GTE_311 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.11")
 GDAL_GTE_3114 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.11.4")
+GDAL_GTE_313 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.13.0")
 
 GEOPANDAS_GTE_10 = version.parse(gpd.__version__) >= version.parse("1.0")
 GEOPANDAS_110 = version.parse(gpd.__version__) == version.parse("1.1.0")
@@ -29,6 +30,7 @@ GEOS_GTE_313 = version.parse(shapely.geos_version_string) >= version.parse("3.13
 
 PANDAS_GTE_20 = version.parse(pd.__version__) >= version.parse("2.0")
 PANDAS_GTE_22 = version.parse(pd.__version__) >= version.parse("2.2")
+PANDAS_GTE_30 = version.parse(pd.__version__) >= version.parse("3.0")
 
 PYOGRIO_GTE_010 = version.parse(pyogrio.__version__) >= version.parse("0.10")
 PYOGRIO_GTE_011 = version.parse(pyogrio.__version__) >= version.parse("0.11")